elan: switch to sigfm
With the introduction of the sigfm algorithm that is able to work
with small sensors, there is no need to treat the Elan fingerprint
sensors as swipe sensors anymore. Switch them to the sigfm algorithm
and make them press type as they are supposed to be. Also Set an enroll
stage count that makes them work reliably.